Replace PriorityDescriptionModal with more generic HelpIconModal
