+++ src/components/index.ts
@@ -96,6 +96,7 @@ export { default as FlowsDeleteButton } from './FlowsDeleteButton.vue'
 export { default as FlowsPageEmptyState } from './FlowsPageEmptyState.vue'
 export { default as FlowsTable } from './FlowsTable.vue'
 export { default as FourZeroFour } from './FourZeroFour.vue'
+export { default as HelpIconModal } from './HelpIconModal.vue'
 export { default as IntervalScheduleForm } from './IntervalScheduleForm.vue'
 export { default as JsonInput } from './JsonInput.vue'
 export { default as JsonView } from './JsonView.vue'
@@ -197,7 +198,6 @@ export { default as WorkerPoolCreateForm } from './WorkerPoolCreateForm.vue'
 export { default as WorkerPoolDetails } from './WorkerPoolDetails.vue'
 export { default as WorkerPoolEditForm } from './WorkerPoolEditForm.vue'
 export { default as WorkerPoolIconText } from './WorkerPoolIconText.vue'
-export { default as WorkersLateIndicator } from './WorkersLateIndicator.vue'
 export { default as WorkerPoolList } from './WorkerPoolList.vue'
 export { default as WorkerPoolMenu } from './WorkerPoolMenu.vue'
 export { default as WorkerPoolQueueCreateForm } from './WorkerPoolQueueCreateForm.vue'
@@ -212,6 +212,7 @@ export { default as WorkerPools } from './WorkerPools.vue'
 export { default as WorkerPoolsPageEmptyState } from './WorkerPoolsPageEmptyState.vue'
 export { default as WorkerPoolToggle } from './WorkerPoolToggle.vue'
 export { default as WorkerPoolTypeSelect } from './WorkerPoolTypeSelect.vue'
+export { default as WorkersLateIndicator } from './WorkersLateIndicator.vue'
 export { default as WorkersTable } from './WorkersTable.vue'
 export { default as WorkQueueCombobox } from './WorkQueueCombobox.vue'
 export { default as WorkQueueCreateForm } from './WorkQueueCreateForm.vue'
@@ -228,4 +229,3 @@ export { default as WorkQueuesTable } from './WorkQueuesTable.vue'
 export { default as WorkQueueStatusBadge } from './WorkQueueStatusBadge.vue'
 export { default as WorkQueueStatusIcon } from './WorkQueueStatusIcon.vue'
 export { default as WorkQueueToggle } from './WorkQueueToggle.vue'
-export { default as PriorityDescriptionModal } from './PriorityDescriptionModal.vue'
